move coordinator into separate file, sensor_descptions to sensor.py
